+++ src/main/java/roomescape/exception/reservation/NotFoundReservationException.java
@@ -1,6 +1,7 @@
 package roomescape.exception.reservation;
 
 import org.springframework.http.HttpStatus;
+
 import roomescape.exception.CustomException;
 
 public class NotFoundReservationException extends CustomException {
