+++ src/main/java/roomescape/web/controller/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.web;
+package roomescape.web.controller;
 
 import java.net.URI;
 import java.util.List;
